refactor(better-local-files): Rename components
