+++ custom-apps/better-local-files/src/components/artists/cards/ArtistCard.tsx
@@ -2,11 +2,11 @@ import React, { useRef } from 'react';
 import styles from '../../../css/app.module.scss';
 import { useIntersectionObserver } from '../../../hooks/use-intersection-observer';
 import { navigateTo } from '../../../helpers/history-helper';
-import { PlayButton } from '../../shared/buttons/play-button';
+import { PlayButton } from '../../shared/buttons/PlayButton';
 import type { Artist } from 'custom-apps/better-local-files/src/models/artist';
-import { MultiTrackMenu } from '../../shared/menus/multi-track-menu';
+import { MultiTrackMenu } from '../../shared/menus/MultiTrackMenu';
 import { ARTIST_ROUTE } from 'custom-apps/better-local-files/src/constants/constants';
-import { TextComponent } from '@shared/components/ui/text/text';
+import { TextComponent } from '@shared/components/ui/TextComponent/TextComponent';
 
 export type Props = {
     artist: Artist;
