+++ custom-apps/better-local-files/src/components/shared/Header.tsx
@@ -1,6 +1,6 @@
 import styles from '../../css/app.module.scss';
 import React from 'react';
-import { TextComponent } from '@shared/components/ui/text/text';
+import { TextComponent } from '@shared/components/ui/TextComponent/TextComponent';
 import useFitText from 'use-fit-text';
 
 export type Props = {
