+++ custom-apps/better-local-files/src/components/shared/filters/SearchInput/SearchInput.module.scss
@@ -1,0 +1,27 @@
+.search-container {
+    display: flex;
+    align-items: center;
+    background-color: rgba(var(--spice-rgb-selected-row), 0.1);
+    border-radius: 4px;
+
+    input {
+        border-style: none;
+        background-color: transparent;
+        color: var(--spice-text) !important;
+    }
+
+    .search-icon {
+        color: var(--spice-text);
+        background-color: transparent;
+        margin-left: 8px;
+        margin-top: 4px;
+        margin-right: 8px;
+        display: inline-block;
+    }
+
+    .clear-icon {
+        color: var(--spice-text);
+        background-color: transparent;
+        display: inline-block;
+    }
+}
