+++ custom-apps/better-local-files/src/components/shared/menus/MenuItemHeading.tsx
@@ -1,5 +1,5 @@
 import React, { type PropsWithChildren } from 'react';
-import { TextComponent } from '@shared/components/ui/text/text';
+import { TextComponent } from '@shared/components/ui/TextComponent/TextComponent';
 
 export function MenuItemHeading(
     props: Readonly<PropsWithChildren>,
