+++ custom-apps/better-local-files/src/components/shared/menus/MultiTrackMenu.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import { SubmenuItem } from './submenu-item';
+import { SubmenuItem } from './SubmenuItem';
 import { getTranslation } from 'custom-apps/better-local-files/src/helpers/translations-helper';
 import type { Track } from 'custom-apps/better-local-files/src/models/track';
-import { PlaylistSelectionMenu } from './playlist-selection-menu';
+import { PlaylistSelectionMenu } from './PlaylistSelectionMenu';
 import { SPOTIFY_MENU_CLASSES } from 'custom-apps/better-local-files/src/constants/constants';
 import { getPlatformApiOrThrow } from '@shared/utils/spicetify-utils';
-import { SpotifyIcon } from '../icons/spotify-icon';
+import { SpotifyIcon } from '../icons/SpotifyIcon';
 import { addToQueuePath } from '../icons/icons';
 import type { PlayerAPI } from '@shared/platform/player';
