+++ custom-apps/better-local-files/src/components/shared/top-bar/TabBarItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import styles from '../../../css/app.module.scss';
+import styles from './TabBar.module.scss';
 import type { TopBarItem } from '../../../models/top-bar-item';
-import { TextComponent } from '@shared/components/ui/text/text';
+import { TextComponent } from '@shared/components/ui/TextComponent/TextComponent';
 
 export type Props = {
     item: TopBarItem;
