+++ custom-apps/better-local-files/src/components/shared/track-list/TrackListGrid.module.scss
@@ -1,0 +1,7 @@
+.display-list {
+    --row-height: 56px;
+}
+
+.display-list-compact {
+    --row-height: 32px;
+}
