+++ custom-apps/better-local-files/src/components/shared/track-list/TrackListRowArtistLink.tsx
@@ -2,7 +2,7 @@ import { ARTIST_ROUTE } from 'custom-apps/better-local-files/src/constants/const
 import { navigateTo } from 'custom-apps/better-local-files/src/helpers/history-helper';
 import type { Track } from 'custom-apps/better-local-files/src/models/track';
 import React from 'react';
-import { TextComponent } from '@shared/components/ui/text/text';
+import { TextComponent } from '@shared/components/ui/TextComponent/TextComponent';
 
 type Props = {
     track: Track;
