+++ custom-apps/better-local-files/src/css/app.module.scss
@@ -1,5 +1,3 @@
-// TODO: Separate in multiple files
-
 // Root
 
 .stretch-container,
@@ -33,40 +31,7 @@
     padding-top: 64px;
 }
 
-.tabBar {
-    ul {
-        display: flex;
-        flex-direction: row;
-        align-items: stretch;
-
-        li {
-            pointer-events: all;
-
-            button {
-                border-radius: 4px;
-                border-style: none;
-                color: var(--spice-text);
-                background-color: transparent;
-                margin: 0 8px 0 0;
-                padding: 8px 16px;
-                cursor: pointer;
-
-                &.active {
-                    background-color: var(--spice-tab-active);
-                    border-radius: 4px;
-                }
-            }
-        }
-
-        select {
-            height: 100%;
-        }
-    }
-}
-
-// Albums
-
-.album-grid {
+.card-grid {
     --minimumColumnWidth: 180px;
     --column-width: minmax(var(--minimumColumnWidth), 1fr);
     --column-count: auto-fill;
@@ -83,7 +48,7 @@
     color: var(--text-subdued);
 }
 
-.album-header {
+.details-page-header {
     align-items: center;
     color: var(--spice-text);
     display: flex;
@@ -94,34 +59,6 @@
         margin-top: 0;
         font-size: 1.5rem;
         font-weight: 700;
-    }
-}
-
-.search-container {
-    display: flex;
-    align-items: center;
-    background-color: rgba(var(--spice-rgb-selected-row), 0.1);
-    border-radius: 4px;
-
-    input {
-        border-style: none;
-        background-color: transparent;
-        color: var(--spice-text) !important;
-    }
-
-    .search-icon {
-        color: var(--spice-text);
-        background-color: transparent;
-        margin-left: 8px;
-        margin-top: 4px;
-        margin-right: 8px;
-        display: inline-block;
-    }
-
-    .clear-icon {
-        color: var(--spice-text);
-        background-color: transparent;
-        display: inline-block;
     }
 }
 
@@ -141,14 +78,6 @@
 .flex-centered {
     display: flex;
     align-items: center;
-}
-
-.display-list {
-    --row-height: 56px;
-}
-
-.display-list-compact {
-    --row-height: 32px;
 }
 
 .limit-lines-2 {
